Fix macro

+++ integration_test_project/macros/cast.sql
@@ -10,7 +10,7 @@
       'databricks': 'STRING',
       'trino': 'VARCHAR'
     },
-    'numeric': {
+    'integer': {
       'postgres': 'NUMERIC',
       'sqlserver': 'FLOAT',
       'bigquery': 'FLOAT64',
